fix studyPage

See merge request s08-webmobile1-sub2/S08P12B304!284

+++ backend/Project/bobs/src/main/java/com/b304/bobs/api/controller/StudyController.java
@@ -1,12 +1,17 @@
 package com.b304.bobs.api.controller;
 
 import com.b304.bobs.api.request.Page.PageReq;
+import com.b304.bobs.api.request.Study.StudyMeetReq;
 import com.b304.bobs.api.request.Study.StudyReq;
+import com.b304.bobs.api.request.Study.StudyUserPageReq;
 import com.b304.bobs.api.response.ModifyRes;
 import com.b304.bobs.api.response.PageRes;
-import com.b304.bobs.api.response.Study.StudyPageRes;
+import com.b304.bobs.api.response.Study.StudyMeetRes;
+import com.b304.bobs.api.response.Study.StudyModifyRes;
 import com.b304.bobs.api.response.Study.StudyRes;
 import com.b304.bobs.api.service.Study.StudyService;
+import com.b304.bobs.api.service.StudyMember.StudyMemberService;
+import com.b304.bobs.db.repository.UserRepository;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
@@ -26,13 +31,57 @@ import java.util.Map;
 public class StudyController {
 
     private final StudyService studyService;
+    private final StudyMemberService studyMemberService;
+    private final UserRepository userRepository;
 
-    @PutMapping("/lock")
-    private ResponseEntity<?> lockStudy(@RequestBody StudyReq studyReq){
+    @PutMapping("/meet")
+    private ResponseEntity<?> meeet(@RequestBody StudyMeetReq studyMeetReq){
         Map<String, Object> map = new HashMap<String, Object>();
 
+        Long user_id = studyMeetReq.getUser_id();
+        Long study_id = studyMeetReq.getStudy_id();
+        boolean study_onair = studyMeetReq.isStudy_onair();
+
         try {
-            ModifyRes modifyRes = studyService.lockStudy(studyReq);
+            if(!studyService.findOneById(study_id).getUser_id().equals(user_id)){
+                map.put("result", false);
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
+            }
+
+            StudyMeetRes studyMeetRes = studyService.studyOnair(studyMeetReq);
+
+            if(studyMeetRes.isResult()){
+                map.put("result", true);
+                map.put("study_id", studyMeetRes.getStudy_id());
+                map.put("study_onair", studyMeetRes.isStudy_onair());
+                return ResponseEntity.status(HttpStatus.OK).body(map);
+            }else{
+                map.put("result", false);
+                map.put("study_id", studyMeetRes.getStudy_id());
+                map.put("study_onair", studyMeetRes.isStudy_onair());
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(map);
+        }
+    }
+
+
+    @PutMapping("/lock")
+    private ResponseEntity<?> lockStudy(@RequestBody StudyUserPageReq studyUserPageReq){
+        Map<String, Object> map = new HashMap<String, Object>();
+        Long user_id = studyUserPageReq.getUser_id();
+        Long study_id = studyUserPageReq.getStudy_id();
+
+        try {
+            if(!studyService.findOneById(study_id).getUser_id().equals(user_id)){
+                map.put("result", false);
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
+            }
+
+            ModifyRes modifyRes = studyService.lockStudy(study_id);
+
             if(modifyRes.getResult()){
                 map.put("study_id", modifyRes.getId());
                 map.put("result", true);
@@ -63,7 +112,6 @@ public class StudyController {
 
             return ResponseEntity.status(HttpStatus.OK).body(map);
 
-
         } catch (Exception e) {
             e.printStackTrace();
             map.put("result", false);
@@ -86,6 +134,24 @@ public class StudyController {
             map.put("result", true);
             return ResponseEntity.status(HttpStatus.OK).body(map);
 
+        } catch (Exception e) {
+            e.printStackTrace();
+            map.put("result", false);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(map);
+        }
+    }
+
+    @PostMapping("/user")
+    public ResponseEntity<Map<String, Object>> getAllByUser(@RequestBody StudyUserPageReq studyUserPageReq) {
+        Map<String, Object> map = new HashMap<>();
+        Long user_id = studyUserPageReq.getUser_id();
+
+        try {
+            PageRes result = studyMemberService.findAllByUser(user_id);
+
+            map.put("data", result.getContents());
+            map.put("result", true);
+            return ResponseEntity.status(HttpStatus.OK).body(map);
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -94,10 +160,10 @@ public class StudyController {
         }
     }
 
-
     @GetMapping("/{studyId}")
     public ResponseEntity<?> getOne(@PathVariable("studyId") Long studyId){
         Map<String, Object> map = new HashMap<String, Object>();
+
         try {
             StudyReq result = studyService.findOneById(studyId);
 
@@ -121,10 +187,10 @@ public class StudyController {
     private ResponseEntity<?> create(@RequestBody StudyReq studyReq){
         Map<String, Object> map = new HashMap<String, Object>();
 
-        if(studyReq.getUser_id() == null) {
-            map.put("result", false);
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
-        }
+//        if(userRepository.isUserExist(studyReq.getUser_id()).isEmpty()) {
+//            map.put("result", false);
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
+//        }
 
         try {
             StudyRes result = studyService.createStudy(studyReq);
@@ -149,27 +215,24 @@ public class StudyController {
     private ResponseEntity<?> modify(@RequestBody StudyReq studyReq){
         Map<String, Object> map = new HashMap<String, Object>();
         Long user_id = studyReq.getUser_id();
+        Long study_id = studyReq.getStudy_id();
 
-        if(user_id == null) {
-            map.put("result", false);
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
-        }
+//        if(userRepository.isUserExist(studyModifyReq.getUser_id()).isEmpty()) {
+//            map.put("result", false);
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
+//        }
         try {
-            StudyReq studyRe = studyService.findOneById(studyReq.getStudy_id());
+            StudyReq studyRes = studyService.findOneById(study_id);
 
-            if(studyRe.getStudy_id() == null || !studyRe.getUser_id().equals(user_id)) {
+            if(!studyRes.getUser_id().equals(user_id)) {
                 map.put("result", false);
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
             }
 
-            StudyPageRes studyPageRes = studyService.modifyStudy(studyReq);
-            System.out.println(studyPageRes.getResult());
-            System.out.println(studyPageRes.getStudyRes());
+            StudyModifyRes studyModifyRes = studyService.modifyStudy(studyReq);
 
-            if(studyPageRes.getResult()){
-                map.put("data", studyPageRes.getStudyRes());
-                map.put("members", studyPageRes.getStudyMemberResList());
-                map.put("member_count", studyPageRes.getMember_count());
+            if(!studyModifyRes.equals(new StudyModifyRes())){
+                map.put("data", studyModifyRes);
                 map.put("result", true);
 
                 return ResponseEntity.status(HttpStatus.OK).body(map);
@@ -184,12 +247,20 @@ public class StudyController {
     }
 
     @DeleteMapping()
-    private ResponseEntity<?> delete(@RequestParam(value="value") Long study_id){
+    private ResponseEntity<?> delete(@RequestBody StudyUserPageReq studyUserPageReq){
         Map<String, Object> map = new HashMap<String, Object>();
+        Long user_id = studyUserPageReq.getUser_id();
+        Long study_id = studyUserPageReq.getStudy_id();
+
         try {
+            if(!studyService.findOneById(study_id).getUser_id().equals(user_id)){
+                map.put("result",false);
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(map);
+            }
+
             ModifyRes modifyRes = studyService.deleteStudy(study_id);
+
             if(modifyRes.getResult()){
-                map.put("study_id", modifyRes.getId());
                 map.put("result",true);
                 return ResponseEntity.status(HttpStatus.OK).body(map);
             }
